move vault-related mutations to their own file
