+++ src/pages/ClaimsPage/useClaimAllocation.test.tsx
@@ -21,7 +21,7 @@ import { useClaimAllocation } from './useClaimAllocation';
 
 let snapshotId: string;
 
-jest.mock('lib/gql/mutations', () => {
+jest.mock('lib/gql/mutations/vaults', () => {
   return {
     addVaultTx: jest.fn().mockReturnValue(Promise.resolve({})),
     addVault: jest
@@ -53,7 +53,6 @@ jest.mock('lib/gql/mutations', () => {
         })
       ),
     savePendingVaultTx: jest.fn(),
-    deletePendingVaultTx: jest.fn(),
   };
 });
